:sparkles: Implement backward compatibility
* v1.0.17 時点での各インタフェースやクラスを復元
