+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Data/DataStore.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: dd8f8fa563954e258a60380c1cbcf3c9
+timeCreated: 1517504412
