+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Data/Entity.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace CAFU.Core.Data {
+
+    public interface IEntity : Entity.IEntity {
+
+    }
+
+    public interface IListEntity<TEntity> : IEntity, IList<TEntity> {
+
+    }
+
+}
