+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Data/Entity.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7e7d0c72c52e4f45b2ec6d4e23cf46ff
+timeCreated: 1517505743
