+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Domain/Model.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace CAFU.Core.Domain {
+
+    public interface IModel : Model.IModel {
+
+    }
+
+    public interface IListModel<TModel> : IModel, IList<TModel> {
+
+    }
+
+}
