+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Domain/Model.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f9350ac9c17444f2bc64ae144e5e2669
+timeCreated: 1517505683
