+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Domain/UseCase.cs
@@ -1,30 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
-using CAFU.Core.Utility;
-
-// ReSharper disable UnusedMember.Global
 
 namespace CAFU.Core.Domain {
 
-    public interface IUseCase {
+    public interface IUseCase : UseCase.IUseCase {
 
     }
 
-    public interface IUseCaseFactory<out TUseCase> where TUseCase : IUseCase {
-
-        TUseCase Create();
-
-    }
-
-    public interface IUseCaseAsSingleton : IUseCase {
-
-    }
-
-    public class DefaultUseCaseFactory<TUseCase> : DefaultFactory<TUseCase>, IUseCaseFactory<TUseCase> where TUseCase : IUseCase, new() {
-
-    }
-
-    public class DefaultSingletonUseCaseFactory<TUseCase> : DefaultSingletonFactory<TUseCase>, IUseCaseFactory<TUseCase> where TUseCase : IUseCaseAsSingleton, new() {
+    public interface IUseCaseAsSingleton : UseCase.IUseCaseAsSingleton, IUseCase {
 
     }
