+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Domain/UseCase.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 30d3f00c3b6e43ae8342174eb646e0ce
+timeCreated: 1517504679
