+++ Assets/Scripts/CAFU/Core/BackwardCompatibility/Presentation/Presenter.cs
@@ -1,31 +1,6 @@
 ﻿using System;
-using CAFU.Core.Utility;
-
-// ReSharper disable UnusedMember.Global
 
 namespace CAFU.Core.Presentation {
-
-    public interface IPresenter {
-
-    }
-
-    public interface IPresenterFactory<out TPresenter> where TPresenter : IPresenter {
-
-        TPresenter Create();
-
-    }
-
-    public class DefaultPresenterFactory<TPresenter> : DefaultFactory<TPresenter>, IPresenterFactory<TPresenter> where TPresenter : IPresenter, new() {
-
-    }
-
-    public static class PresenterExtension {
-
-        public static TPresenter As<TPresenter>(this IPresenter presenter) where TPresenter : class, IPresenter {
-            return presenter as TPresenter;
-        }
-
-    }
 
     [Obsolete("Please use IPresenterFactory<TPresenter> instead of this interface.")]
     public interface IPresenterBuilder {
