+++ Assets/Scripts/CAFU/Core/Data/DataStore.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 52fba38b02274914bc7b9294e42b8284
+timeCreated: 1517505490
