+++ Assets/Scripts/CAFU/Core/Data/DataStore/DataStore.cs
@@ -2,6 +2,7 @@
 using System.Text.RegularExpressions;
 using CAFU.Core.Utility;
 using UnityEngine;
+
 // ReSharper disable UnusedMember.Global
 
 namespace CAFU.Core.Data {
