+++ Assets/Scripts/CAFU/Core/Data/Entity.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 60e6ac00adf9402e903e533e161b6a35
+timeCreated: 1517505496
