+++ Assets/Scripts/CAFU/Core/Data/Entity/Entity.cs.meta
