+++ Assets/Scripts/CAFU/Core/Domain/Model.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3744acbb83014fb0a9f4dd659bc63d5c
+timeCreated: 1517505438
