+++ Assets/Scripts/CAFU/Core/Domain/Model/Model.cs.meta
