+++ Assets/Scripts/CAFU/Core/Domain/Translator.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 33f77104791042aaa32b59ceab763688
+timeCreated: 1517505455
