+++ Assets/Scripts/CAFU/Core/Domain/Translator/Translator.cs
