+++ Assets/Scripts/CAFU/Core/Domain/UseCase.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 302bfba485ce4a7da85b0a498d531128
+timeCreated: 1517505462
