+++ Assets/Scripts/CAFU/Core/Domain/UseCase/UseCase.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using CAFU.Core.Utility;
+
+// ReSharper disable UnusedMember.Global
+
+namespace CAFU.Core.Domain.UseCase {
+
+    public interface IUseCase {
+
+    }
+
+    public interface IUseCaseFactory<out TUseCase> where TUseCase : IUseCase {
+
+        TUseCase Create();
+
+    }
+
+    public interface IUseCaseAsSingleton : IUseCase {
+
+    }
+
+    public class DefaultUseCaseFactory<TUseCase> : DefaultFactory<TUseCase>, IUseCaseFactory<TUseCase> where TUseCase : IUseCase, new() {
+
+    }
+
+    public class DefaultSingletonUseCaseFactory<TUseCase> : DefaultSingletonFactory<TUseCase>, IUseCaseFactory<TUseCase> where TUseCase : IUseCaseAsSingleton, new() {
+
+    }
+
+}
