+++ Assets/Scripts/CAFU/Core/Domain/UseCase/UseCase.cs.meta
