+++ Assets/Scripts/CAFU/Core/Presentation/View/Controller.cs
@@ -1,38 +1,28 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UniRx;
+
 // ReSharper disable VirtualMemberNeverOverridden.Global
 // ReSharper disable MemberCanBePrivate.Global
 // ReSharper disable UnusedMember.Global
 
 namespace CAFU.Core.Presentation.View {
 
-    public interface IController {
+    public interface IController : IView {
 
         IPresenter Presenter { get; set; }
 
     }
 
-    public abstract class Controller<TPresenter> : ObservableLifecycleMonoBehaviour, IController
-        where TPresenter : IPresenter, new() {
-
-        private static IPresenterFactory<TPresenter> presenterFactory = new DefaultPresenterFactory<TPresenter>();
-
-        public static IPresenterFactory<TPresenter> PresenterFactory {
-            get {
-                return presenterFactory;
-            }
-            set {
-                presenterFactory = value;
-            }
-        }
+    public abstract class Controller<TPresenter, TPresenterFactory> : ObservableLifecycleMonoBehaviour, IController
+        where TPresenter : IPresenter, new()
+        where TPresenterFactory : IPresenterFactory<TPresenter>, new() {
 
         IPresenter IController.Presenter { get; set; }
 
         protected override void OnAwake() {
             base.OnAwake();
             ControllerInstanceManager.Instance.Register(this);
-            ((IController)this).Presenter = PresenterFactory.Create();
+            ((IController)this).Presenter = new TPresenterFactory().Create();
         }
 
         protected override void OnDestroy() {
@@ -87,40 +77,6 @@ namespace CAFU.Core.Presentation.View {
         public static TController As<TController>(this IController controller) where TController : class, IController {
             return controller as TController;
         }
-
-    }
-
-    [Obsolete("Please use Zenject instead of this.")]
-    public interface IControllerBuilder {
-
-        void Build();
-
-    }
-
-    [Obsolete("Please use Controller<TPresenter> instead of this.")]
-    public abstract class ControllerAsSingleton<TViewController, TPresenter> : Controller<TPresenter>
-        where TViewController : ControllerAsSingleton<TViewController, TPresenter>
-        where TPresenter : IPresenter, new() {
-
-        // ReSharper disable once UnusedAutoPropertyAccessor.Global
-        public static TViewController Instance { get; private set; }
-
-        protected override void OnAwake() {
-            base.OnAwake();
-            Instance = (TViewController)this;
-        }
-
-        protected override void OnDestroy() {
-            Instance = null;
-        }
-
-    }
-
-    // コンパイルエラー回避のために残してあります
-    [Obsolete("Please use Controller<TPresenter> instead of this.")]
-    public abstract class ViewControllerBaseAsSingleton<TViewController, TPresenter> : ControllerAsSingleton<TViewController, TPresenter>
-        where TViewController : ControllerAsSingleton<TViewController, TPresenter>
-        where TPresenter : IPresenter, new() {
 
     }
